rename package (typo)
